SQLite: Check filename before deleting (GHSA-6vfx-w7q3-42gq)

+++ admin/drivers/sqlite.inc.php
@@ -496,6 +496,9 @@ if (isset($_GET["sqlite"])) {
 	{
 		Connection::get()->close(); // to unlock file, doesn't work in PDO on Windows
 		foreach ($databases as $db) {
+			if (!check_sqlite_name($db)) {
+				return false;
+			}
 			if (!@unlink($db)) {
 				Connection::get()->setError(lang('File exists.'));
 				return false;
